split old code into more components: taskItem, AddTaskForm. Button to delete tasks, possibility to mark task individually as complete/incomplete
