+++ src/App.jsx
@@ -1,14 +1,13 @@
 
+import { HeaderSection } from "./components/HeaderSection"
+import { AddTaskForm } from "./components/AddTaskForm"
 import { TaskSection } from "./components/TaskSection"
-import { HeaderSection
- } from "./components/HeaderSection"
-import { AddTaskSection } from "./components/AddTaskSection"
 
 export const App = () => {
   return (
     <>
       <HeaderSection />
-      {/* <AddTaskSection /> */}
+      <AddTaskForm />
       <TaskSection />
     </>
   )
