--- src/components/AddTaskSection.jsx
@@ -1,10 +1,0 @@
-export const AddTaskSection = () => {
-  return (
-    <section>
-      <form action="">
-        <input type="text" />
-        <button type="submit">Add task</button>
-      </form>
-    </section>
-  )
-}
