+++ src/stores/useTodoStore.jsx
@@ -2,25 +2,20 @@ import { create } from 'zustand'
 
 export const useTodoStore = create((set) => ({
   // CREATE STATE = what is in the store (initial state): Array of tasks, if task is completed, remove task 
-  task: [ ],
-  isCompleted: false,
+  task: [],
+  // isCompleted: false,
   removeTask: false,
 
   // SET STATE = change what is in the store
   // Add task
-  addTask: (newTask) => set((state) => ({
-    task: [...state.task, { newTask, isCompleted: false}]
-  })),
-
-  // Toggle between completed/not completed
-  toggleIsCompleted: () => set((state) => ({
-    isCompleted: !state.isCompleted
+  addTask: (newTask) => set((state) => ({ task: [...state.task, { newTask, isCompleted: false, id: Date.now()}]
   })),
 
   // Remove task
-  // removeTask: () => set((state) => ({
-  //   removeTask: state.task.remove
-  // }))
+  removeTask: (id) => set((state) => ({ task: state.task.filter(task => task.id !== id) })),
+
+  // Toggle between completed/not completed
+  toggleIsCompleted: (id) => set((state) => ({ task: state.task.map(task => task.id === id ? {...task, isCompleted: !task.isCompleted}: task)})),
  
 }))
